Move MetricNumeric and isFiniteMetricValue into Numeric.swift
Identifiers.swift is for identifier and attribute types; the numeric metric
vocabulary is unrelated to it. Pure move, no behavior change.
